add my-offering page
